requirements.txt
django>=1.7.0
qrcode
django-allauth
django-otp

